+++ src/data/kpis.js
@@ -1,4 +1,5 @@
-import { getBtcHistoricalData } from '../services/btcDataService.js'
 import { createKpisFromHistoricalData } from '../utils/kpiCalculations.js'
 
-export const kpis = createKpisFromHistoricalData(getBtcHistoricalData())
+export function createKpis(data) {
+  return createKpisFromHistoricalData(data)
+}